Remove configurations keyword from internal code
The app.config was always extracting the "organizations" key,
but there is no point in doing that as it is the only primary key.
Thus we have removed that.

On the other hand, we keep it for the API as it helps to understand
the data needed from the user perspective.